Update README.md
The project is an ATM Management System developed in C++. It allows users to perform essential banking operations such as checking account balance, withdrawing money, depositing money, and exiting the system. The system initializes with predefined account holders and assigns random balances to each. Users can securely log in using their account name and a PIN, and interact with an intuitive menu to perform transactions.